refac : 도메인으로 폴더 명 변경
1. Post를 도메인으로 하고 싶다.
1. Profile을 embedable로 하고 싶다.
