+++ src/main/java/com/fromapril/member/domain/member/Profile.java
@@ -1,9 +1,8 @@
-package com.fromapril.member.model.member;
+package com.fromapril.member.domain.member;
 
 import jakarta.persistence.*;
 import lombok.Data;
 import lombok.NoArgsConstructor;
-import lombok.RequiredArgsConstructor;
 
 @NoArgsConstructor
 @Data
