+++ src/main/java/com/fromapril/member/domain/point/PointSaveStaus.java
@@ -1,4 +1,4 @@
-package com.fromapril.member.model.point;
+package com.fromapril.member.domain.point;
 
 public enum PointSaveStaus {
     NOT_USED,
